% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_pangoro}
\alias{collapse_pangoro}
\title{Collapse PANGO lineage names}
\usage{
collapse_pangoro(pangoro, input, max_level = NULL, simplify = TRUE)
}
\arguments{
\item{pangoro}{Pangoro object.}

\item{input}{Vector of lineage names to collapse (character vector).}

\item{max_level}{How far to expand the lineage name (default: maximum).}

\item{simplify}{Boolean, pass to \code{\link{mapply}}.}
}
\description{
Collapse provided PANGO lineage names completely or up to a maximum depth.
Recombinant variants will not be collapsed. NA_character values will be carried through.
}
\examples{
\dontrun{
my_pangoro <- pangoro()
x <- c('B.1.1.529.1', 'B.1.1.529.2.75.1.2', 'BA.1', NA_character_)
collapse_pangoro(my_pangoro, x)
collapse_pangoro(my_pangoro, x, 1)
collapse_pangoro(my_pangoro, x, 2)
}
}
