% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_pangoro}
\alias{expand_pangoro}
\title{Expand lineage from alias}
\usage{
expand_pangoro(pangoro, input = character(), max_level = NULL, simplify = TRUE)
}
\arguments{
\item{pangoro}{Pangoro object.}

\item{input}{Vector of lineage names to expand (character vector).}

\item{max_level}{How far to expand the lineage name (default: maximum).}

\item{simplify}{Boolean, pass to \code{\link{mapply}}}
}
\description{
Expand lineage names from an alias to parent name up to a maximum step.
Recombinant variants will not be expanded. NA_character values will be carried through.
}
\details{
Although several string splits occur, this function can expand 1e6 values in <2 minutes.
}
\examples{
\dontrun{
my_pangoro <- pangoro()
expand_pangoro(my_pangoro, c('BA.1', 'BA.1.2.3', 'BL.2'))
expand_pangoro(my_pangoro, c('BA.1', 'BA.1.2.3', 'BL.2'), 1)
expand_pangoro(my_pangoro, c('BA.1', 'BA.1.2.3', 'BL.2'), 2)
expand_pangoro(my_pangoro, c('BA.1', 'BA.1.2.3', 'BL.2'), 200) # Only goes as far as it can
}
}
