% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_pango_alias}
\alias{fetch_pango_alias}
\title{Fetch alias table from PANGO}
\usage{
fetch_pango_alias(
  url =
    "https://raw.githubusercontent.com/cov-lineages/pango-designation/master/pango_designation/alias_key.json"
)
}
\arguments{
\item{url}{Character value defining the raw data location for the alias table.}
}
\description{
Pulls JSON and formats into a flat table. Recombinants determined by the number
of lineages associated with alias.
}
\examples{
\dontrun{
fetch_pango_alias()
}

}
