% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_pango_full_tbl}
\alias{fetch_pango_full_tbl}
\title{Fetch full summary table from PANGO.}
\usage{
fetch_pango_full_tbl(
  url =
    "https://github.com/cov-lineages/lineages-website/blob/master/_data/lineage_data.full.json?raw=true",
  pattern = "^([Aa]lias of)\\\\s?([A-Z\\\\.\\\\d]*)[,]?\\\\s?.*$",
  description_col = "Description"
)
}
\arguments{
\item{url}{Character value defining the location of the JSON file containing the complete summary details from PANGOLIN.}

\item{pattern}{Regular expression to search description column for alias details.}

\item{description_col}{Name of the description column (default: 'Description').}
}
\description{
Pulls the lineage notes data from the PANGO designation GitHub: \url{https://github.com/cov-lineages/lineages-website/blob/master/_data/lineage_data.full.json?raw=true}.
Compared to \code{\link{fetch_pango_summary_tbl}}, this function pulls the entire set of summary information available from PANGO JSON file.
To provide the full lineage name, the *Description* column is searched for alias information based upon
the regular expression passed to the \code{pattern} parameter. If this parameter is set to \code{NULL},
the alias search will be ignored and the \code{full_name} column will be excluded.
}
\examples{
\dontrun{
fetch_pango_full_tbl()
fetch_pango_full_tbl(pattern = NULL)
}
}
