% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_pango_notes}
\alias{fetch_pango_notes}
\title{Fetch PANGO lineage notes}
\usage{
fetch_pango_notes(
  url =
    "https://raw.githubusercontent.com/cov-lineages/pango-designation/master/lineage_notes.txt",
  pattern = "^([Aa]lias of)\\\\s?([A-Z\\\\.\\\\d]*)[,]?\\\\s?.*$",
  description_col = "Description"
)
}
\arguments{
\item{url}{Character value defining the location of the lineage notes from PANGOLIN.}

\item{pattern}{Regular expression to search description column for alias details.}

\item{description_col}{Name of the description column (default: 'Description').}
}
\description{
Pulls the lineage notes data from the PANGO designation GitHub: \url{https://raw.githubusercontent.com/cov-lineages/pango-designation}.
To provide the full lineage name, the *Description* column is searched for alias information based upon
the regular expression passed to the \code{pattern} parameter. If this parameter is set to \code{NULL},
the alias search will be ignored and the \code{full_name} column will be excluded.
}
\examples{
\dontrun{
fetch_pango_notes()
fetch_pango_notes(pattern = NULL)
}
}
