% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_pango_summary_tbl}
\alias{fetch_pango_summary_tbl}
\title{Fetch summary table from PANGO.}
\usage{
fetch_pango_summary_tbl(
  url = "https://cov-lineages.org/lineage_list.html",
  pattern = "^([Aa]lias of)\\\\s?([A-Z\\\\.\\\\d]*)[,]?\\\\s?.*$",
  description_col = "Description",
  slice = 1,
  ...
)
}
\arguments{
\item{url}{Character value defining the location of the summary table from PANGOLIN.}

\item{pattern}{Regular expression to search description column for alias details.}

\item{description_col}{Name of the description column (default: 'Description').}

\item{slice}{Integer value (default: 1), controls which table from the page to extract.}

\item{...}{Additional parameters to send to \code{\link[rvest]{read_html}}.}
}
\description{
Pulls what is available on the COVID-19 lineage website: \url{https://cov-lineages.org/lineage_list.html}.
As of the recent changes in late 2022, this website only has a annual subset of lineages. Since
the table is embedded within the web-page the \code{rvest} package is used to read the HTML content.
To provide the full lineage name, the *Description* column is searched for Alias information based upon
the regular expression passed to the \code{pattern} parameter. If this parameter is set to \code{NULL},
the alias search will be ignored and the \code{full_name} column will be excluded.
}
\examples{
\dontrun{
fetch_pango_summary_tbl()
}
}
