% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_recombinant}
\alias{is_recombinant}
\title{Detect recombinants}
\usage{
is_recombinant(pangoro, input, use_names = FALSE)
}
\arguments{
\item{pangoro}{Pangoro object.}

\item{input}{Vector of lineage names to expand (character vector).}

\item{use_names}{Boolean to determine if logical or character value of recombinant names used.}
}
\value{
Either boolean or character vector
}
\description{
Determine if any of the input lineages expand back to a recombinant node.
}
\examples{
\dontrun{
my_pangoro <- pangoRo::pangoro()
is_recombinant(my_pangoro, c('EG.1', 'EC.1'))
}
}
