% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_children}
\alias{list_children}
\title{List children of provided alias}
\usage{
list_children(pangoro, input)
}
\arguments{
\item{pangoro}{Pangoro S3 class object.}

\item{input}{Lineage name (scalar).}
}
\description{
Provide a lineage and return a vector of the child aliases.
}
\examples{
\dontrun{
my_pangoro <- pangoro()
list_children(my_pangoro, 'BL.1')
lapply(c('BL.1', 'BA.1', 'BE.9.1', 'BQ.4'), list_children, pangoro = my_pangoro)
}
}
