% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_parents}
\alias{list_parents}
\title{List parents of provided alias}
\usage{
list_parents(pangoro, input)
}
\arguments{
\item{pangoro}{Pangoro S3 class object.}

\item{input}{Lineage name (scalar).}
}
\description{
Provide a lineage and return a vector of the parent aliases. Not vectorized.
}
\examples{
\dontrun{
my_pangoro <- pangoro()
list_parents(my_pangoro, 'BL.1')
lapply(c('BL.1', 'BA.1', 'BE.9.1', 'BQ.4', 'XBA.1'), list_parents, pangoro = my_pangoro)
}
}
