% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{pangoro}
\alias{pangoro}
\title{Create pangoro object}
\usage{
pangoro(
  path =
    "https://raw.githubusercontent.com/cov-lineages/pango-designation/master/pango_designation/alias_key.json",
  refresh = FALSE,
  offline = FALSE,
  rm_spec_char = TRUE
)
}
\arguments{
\item{path}{Path to PANGO lineage JSON file.}

\item{refresh}{Boolean value to force cache to refresh (default: \code{FALSE})}

\item{offline}{Boolean value, if \code{TRUE} use a saved version of alias table within package.}

\item{rm_spec_char}{Boolean (default: \code{TRUE})}
}
\description{
Create S3 class object that will contain the necessary PANGO alias details for
working with other functions (expanding and collapsing names). Will cache the PANGO lineage table for
the current R session.
}
\examples{
\dontrun{
my_pangoro <- pangoro() # First time will fetch from web
my_pangoro <- pangoro() # Second time will fetch from cache
my_pangoro <- pangoro(refresh = TRUE) # Force cache to refresh
}
}
