% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.pangoro}
\alias{print.pangoro}
\title{Print pangoRo object}
\usage{
\method{print}{pangoro}(x, ...)
}
\arguments{
\item{x}{pangoRo object.}

\item{...}{Not used.}
}
\description{
Print the alias and lineage table from the pangoRo object.
}
