% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_pangoro}
\alias{search_pangoro}
\title{Search for alias within PANGO lineages}
\usage{
search_pangoro(pangoro, input, search, direction = "both")
}
\arguments{
\item{pangoro}{Pangoro S3 class object.}

\item{input}{Lineage name (scalar).}

\item{search}{Alias name to search if related to input list.}

\item{direction}{Search parents (\code{up}) and/or children (\code{down}) of alias (default: \code{both}).}
}
\description{
Search for alias within PANGO lineages
}
\examples{
\dontrun{
my_pangoro <- pangoro()
search_pangoro(my_pangoro, 'BA.5', 'BA.5.1')
mapply(c('BA.5', 'BA.5'),c('BA.5.1', 'BL.1'), FUN = function(x,y) {search_pangoro(my_pangoro,x,y)})
}

}
