% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_pangoro}
\alias{sort_pangoro}
\title{Sort PANGO lineages}
\usage{
sort_pangoro(pangoro, x, index_only = FALSE, na.last = FALSE, ...)
}
\arguments{
\item{pangoro}{Pangoro class object}

\item{x}{Vector of pango lineages}

\item{index_only}{Boolean value, to return index or sorted values (default: \code{FALSE}).}

\item{na.last}{Sort NAs top or bottom (default: TRUE, sort to top)}

\item{...}{Additional parameters to \code{\link{order}}.}
}
\description{
Return ordering for PANGO lineages based upon a provided list of names which
may include aliases.
}
\examples{
\dontrun{
my_pangoro <- pangoro()
x <- c('BL.2', 'BA.1', 'BA.1.2.3', 'BA.1.223.3', 'BA.1.4.3','BL.2', 'BA.1')
sort_pangoro(my_pangoro, x)
sort_pangoro(my_pangoro, x, index_only = TRUE)
sort_pangoro(my_pangoro, x, na.last = TRUE)
sort(x) # Compare to normal sort in base R
}
}
