% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.pangoro}
\alias{summary.pangoro}
\title{Summarize pangoRo object}
\usage{
\method{summary}{pangoro}(object, ...)
}
\arguments{
\item{object}{pangoRo object.}

\item{...}{Not used.}
}
\description{
Supply basic summary details of the pangoRo object.
}
